chore: move ModelUploader component
