+++ client/src/features/model/components/index.ts
@@ -13,5 +13,6 @@ export { ModelInferenceTab } from "./tabs/ModelInferenceTab";
 export { NeuralNetworkVisualization } from "./tabs/NeuralNetworkVisualization";
 
 // upload steps
+export { ModelUploader } from './upload-steps/ModelUploader';
 export { StepHeader } from './upload-steps/StepHeader';
 export { UploadButton } from './upload-steps/UploadButton';
